+++ novaideo/__init__.py
@@ -1218,6 +1218,7 @@ def main(global_config, **settings):
     config.add_translation_dirs('deform:locale/')
     config.add_translation_dirs('colander:locale/')
     config.include('.graphql')
+    config.include('.oidc_sso')
     config.include("pyramid_sms")
     config.scan()
     config.add_static_view('novaideostatic',